extract LibraryComponent into AppComponent
